Use standard ping.json from django-moj-irat
[django-moj-irat](https://github.com/ministryofjustice/django-moj-irat)
offers a ready-made, well-tested `ping.json` implementation that
behaves in the way we want (using environment variables to provide its
values), so we should use that instead of maintaining our own.

This depends on ministryofjustice/django-moj-irat#4, which adds support
for Python 2.7.

For this to behave as intended, we need `APP_BUILD_DATE`,
`APP_GIT_COMMIT`, `APP_BUILD_TAG`, and `APP_VERSION` environment
variables to be appropriately set.