Reimplement healthchecks using moj_irat helpers
As with `PingJsonView` we can use `moj_irat`'s `HealthcheckView` (and
related helpers) to drastically reduce the amount of logic we have to
implement to monitor our upstream dependencies.

We get database checking for free, by default, and we can trivially
monitor URLs (and their contents!) using `UrlHealthcheck`. These checks
should probably live in other apps (and the healthcheck app should go
away), in time, but that time is not now.

It's also of note that this doesn't yet reimplement the `s3_courts_data`
healthcheck: that seems complex enough to warrant its own commit, which
will come shortly.