Reimplement data freshness healthcheck
This re-adds a healthcheck, using the new `moj_irat` framework to warn
if the current data in the system is both different to the latest
version on the remote server and older than 10 minutes; if that's the
case, it would suggest something is wrong with the ingestion.

This is also much more thoroughly tested in an attempt to more clearly
document the expected behaviour.

There's a slight bastardisation of `UrlHealthcheck.success_response`
here: we're making it return a failure response if the data is stale,
but I couldn't see a better way to do this without having to
significantly reimplement `UrlHealthcheck.__call__` in
`CourtDataFreshnessHealthcheck`.