Implement multiple database support for courts.json data ingestion
Currently, calling ingest will always load the courts data into the default
database. When importing data, we want to import it into a temporary
database, and then swap this with the live one.
This change allows the --database parameter to be passed to the
populate-db script, and so the Ingest, which uses the identifier
to get/create/save objects to the matching database entry in the
django settings file.

(Closes #122)