Increase timeout for postcodeinfo calls
When we search for a partial postcode, and the postcodeinfo service has not
already cached the result, the time taken to fetch the result is greater than
the default 5 second timeout for the postcodeinfo client. This changes increases
the timeout to 30 seconds, which is greater than the observed maximum partial
search query time of ~25 seconds. The timeout can also be specified using the
POSTCODEINFO_API_TIMEOUT environment variable.