Add a marker file that indicates whether the last data import was successful
Ingestion runs in two stages, loading the remote courts data into the application,
and then importing it into the database. If the first stage is successful but the
second fails, there will only be one warning then the application will stop warning.
This change ensures that it retries and also keeps warning.