Put leaflets behind a feature flag
Before this change the leaflets markup had been commented out in the
template but the tests had not been commented out so some tests were
failing. This change introduces a feature flag that is switched off and
puts the markup behind that feature flag.

The tests now explicitly enable the feature flag and there are two new
tests verifying that the feature flag properly hides the leaflets
section.

When we come to enable leaflets we can now just enable the flag
in the environment we're interested in and then in production.