Simplify and fix Dockerfile
Backstory:
We noticed that mgirations are not being run when an app is deployed.
Investigating pointed us towards commit 8496ce1 which uses a `run.sh`
that doesn't respect the `DOCKER_STATE` environment variable.

I haven't been able to tie this back to a pull request. It looks like
there was a large history change at some point in the past that this was
involved in.

Further investigation revealed that the Dockerfile seems to be more
complicated than is necesarry.

Solution:
This change strips down and fixes the Dockerfile in a number of ways:
- Remove pip install uWSGI as the dependency is already in
  requirements.txt
- Simplify how the code is copied into place.
- Use the main `run.sh` which includes the `DOCKER_STATE` commands.
- Generate assets in place rather than doing so in a site directory
  (`/srv/additional_files`)
And outside of the Docker file:
- Remove cds from `setup_npm.sh` as the assets are no longer built on
  the side.
- Remove `docker/run.sh` as it confuses the situation.