Simplify postcode search
The intention of this change is to simplify and generally improve the
postcode search.

The previous solution uses a rather complex raw query and an even more
complex deduping approach (apparently for some marginal performance
improvements). Given the effort placed on performance here I have
assumed that there can be quite a lot of courts returned. I've tried to
address that in a few of ways.

1. Replace the raw SQL with a Django ORM based query
This means that we can more easily join through the courts table to get
the areas of law. This reduces the number of queries from N+2 to 1. This
is likely to be the biggest gain. Unfortunately it still needs the raw
where clause (see: http://stackoverflow.com/a/11906048/15720)

2. Move the area of law check into the query
We are querying through those join points anyway so we may as well do
it all in the database.

3. Move the deduping into the query
The database is already handling this data and it is more obvious what's
happening.

4. Remove the sorting
The only place that this method is used subsequently sorts the results by
proximity so there's no need to sort.