Remove staging Django settings file
This file is not used when deployed to staging. Our deployment
infrastructure expects all configuration to be handled by environment
variables. This change moves us back towards that.

One thing that's added to help with that is a `is_enabled` helper for
feature flags. I have just added into `settings/base.py` for now as the
settings files are the only place we expect to use it.