Remove pingdom search statistics healthcheck
Upgrading Django highlighted that the `SearchStatistic` model does not
exist (it looks like it's been purged from the history) and therefore
this healthcheck now throws a 500. Kibana shows that this healthcheck
has not been requested in the past few weeks so I'm going to assume it's
no longer in use.