Pin npm dependency chalk to 1.1.3
Chalk is being pulled in as an unrestricted dependency. It looks like
Chalk version 2.0.0 went to ES6 which now causes it to break on our
build. This change pins it to the most recent version before 2.0.0.

It also checks that /usr/bin/node doesn't exist before trying to symlink
it so that `setup_npm.sh` can be used more easily from vagrant.