Update docker image
There are a few major changes here.

Changed base image to python:2.7 as this then handles a lot of the
python dependency stuff we were doing. This changes OS from
ubuntu:14.0.4 to debian:jessie

Move `COPY . .` much further down the file and introduce more specific
`COPY` commands through the dockerfile to allow for better use of the
cache and quicker build times.

Remove python and pinned postgis dependencies from `apt/base.sh` as they
are not needed with the python base image.

Fix the postgres container for easier local testing.

Make the `./run.sh` executable so that it can called directly in the
`COMMAND`.

Be explicit about the port we are exposing.