Switch to a slim base image
The standard python base image is almost 700MB. This change reduces the
size of the built image by using the slim base image and adding just the
dependencies we need.

There are a number of other changes to the Dockerfile to reduce the size
of the image.
Combine the `apt/production.sh` and asset building stages so that the
asset building dependencies can be removed without ever getting commited
to a layer.
Remove the `node_modules`, asset building apt dependencies and apt lists
in the system dependencies / asset building stage.
Do not use the cache when installing python dependencies.