Fix Dockerfile pre-1.9 versions of docker
The docker version on the build environment is pre-1.9 and therefore
does not understand the `ARG` command Dockerfiles. This change is not
ideal because it means changes to the Dockerfile now also need to be
applied to the Dockerfile.dev but upgrading the version of docker in the
build environment is a slightly bigger task and this is blocking
deploys.

https://docs.docker.com/release-notes/docker-engine/#190-2015-11-03